Move capability-specific asserts out of the core CRUD contract
The core CRUD contract reached into two optional capabilities: the
queue-wait percentile assertions used the analytics view, and the
recurring-master round-trip used the recurring view. On a store without
those capabilities the whole CRUD contract degraded to not-applicable,
and the percentile semantics had no home outside CRUD at all.

Move the percentile cases to the analytics contract and the
recurring-master child round-trip to the recurring contract, where each
accessor is correctly gated. The recurring contract gains a small hook
to reach the core store for the child create and read-back.
